REALMC-11731: add template literals and more (#49)
- Prefer the exported value when doing type conversions to preserve wra
- Support conversion from string types to []byte.
- Fixed dynamic variable resolution in iteration scopes
- Implemented template literals and \u{xxxx}.
- Support chained template literals.
- Fixed the use of outer scope 'this' in arrow functions.
- Support 0o and 0b number literals.
- Fixed ExportType() for Date objects
- Try to fix sourcemaps with urls

+++ value.go
@@ -45,7 +45,7 @@ var (
 	reflectTypeInt64  = reflect.TypeOf(int64(0))
 	reflectTypeInt32  = reflect.TypeOf(int32(0))
 	reflectTypeUInt32 = reflect.TypeOf(uint32(0))
-	reflectTypeInt    = reflect.TypeOf(0)
+	reflectTypeInt    = reflect.TypeOf(int64(0))
 	reflectTypeBool   = reflect.TypeOf(false)
 	reflectTypeNil    = reflect.TypeOf(nil)
 	reflectTypeFloat  = reflect.TypeOf(float64(0))
